Refactor code for event listeners in controllers instead of models
